requirements.txt
Django==1.7
djangorestframework==3.1.3
markdown==2.1.0
django-filter==0.10.0
envvars==0.3.0
django-rest-swagger